refactoring(foe-data): Rearrange defaultCost
