+++ lib/foe-data/ages-cost/HighMiddleAges.js
@@ -1,4 +1,5 @@
-const { defaultCost, generateReward } = require("../utils");
+const { generateReward } = require("../utils");
+const defaultCost = require("./defaultCost");
 
 module.exports = [
   { cost: 50, reward: generateReward(5) },
